move tests, and change getDictionary behaviour
